statustimers: small refactor.
