+++ runelite-client/src/main/java/net/runelite/client/plugins/statustimers/Victim.java
@@ -6,6 +6,10 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.Setter;
 import net.runelite.api.Actor;
+import net.runelite.api.HeadIcon;
+import net.runelite.api.NPC;
+import net.runelite.api.NPCDefinition;
+import net.runelite.api.Player;
 import net.runelite.api.Prayer;
 
 @Getter(AccessLevel.PACKAGE)
@@ -33,6 +37,14 @@ class Victim
 		this.overlaysDrawn = 0;
 		this.timerMap = new HashMap<>();
 		this.immunityMap = new HashMap<>();
+	}
+
+	void update()
+	{
+		setLastPrayer(getCurrentPrayer());
+		setSpotAnimLastTick(getSpotAnim());
+		setCurrentPrayer(updatePrayer(getActor()));
+		setSpotAnim(getActor().getSpotAnimation());
 	}
 
 	void timerToImmunity(int ticks)
@@ -73,6 +85,55 @@ class Victim
 			{
 				getImmunityMap().remove(type);
 			}
+		}
+	}
+
+	private Prayer updatePrayer(Actor actor)
+	{
+		if (actor instanceof NPC)
+		{
+			final NPCDefinition def = ((NPC) actor).getDefinition();
+			if (def == null)
+			{
+				return null;
+			}
+			final HeadIcon icon = def.getOverheadIcon();
+			if (icon == null)
+			{
+				return null;
+			}
+			return headIconToPrayer(def.getOverheadIcon());
+		}
+		else if (actor instanceof Player)
+		{
+			final HeadIcon icon = ((Player) actor).getOverheadIcon();
+			if (icon == null)
+			{
+				return null;
+			}
+			return headIconToPrayer(icon);
+		}
+		return null;
+	}
+
+	private Prayer headIconToPrayer(HeadIcon icon)
+	{
+		switch (icon)
+		{
+			case MELEE:
+				return Prayer.PROTECT_FROM_MELEE;
+			case RANGED:
+				return Prayer.PROTECT_FROM_MISSILES;
+			case MAGIC:
+				return Prayer.PROTECT_FROM_MAGIC;
+			case RETRIBUTION:
+				return Prayer.RETRIBUTION;
+			case SMITE:
+				return Prayer.SMITE;
+			case REDEMPTION:
+				return Prayer.REDEMPTION;
+			default:
+				return null;
 		}
 	}
 }
